pyproject.toml
[tool.poetry]
name = "Git bark"
version = "0.1.0"
description = ""
authors = ["Elias Bonnici <elias.bonnici@yubico.com>"]
readme = "README.md"
packages=[{include = "gitbark"}]

[tool.poetry.dependencies]
python = "^3.9"
pgpy = "^0.6.0"
pyyaml = "^6.0"
click = "^8.1.3"
colorama = "^0.4.6"
pygit2 = "^1.12.1"

[tool.poetry.scripts]
bark = "gitbark._cli.__main__:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]